move existing 'register markdown handler' helper to live alongside other common markdown code
