+++ lib/cdo/pegasus/actionview_sinatra.rb
@@ -1,5 +1,5 @@
 require 'action_view'
-require 'cdo/markdown_handler'
+require 'cdo/markdown/handler'
 require 'cdo/markdown/renderer'
 require 'cdo/redcarpet/inline'
 require 'haml'
@@ -13,9 +13,9 @@ module ActionViewSinatra
       @sinatra = sinatra
       super(*args)
 
-      MarkdownHandler.register(Cdo::Markdown::Renderer, Cdo::Markdown::Renderer::OPTIONS)
-      MarkdownHandler.register_html_safe(Redcarpet::Render::Safe)
-      MarkdownHandler.register_inline(Redcarpet::Render::Inline.new({filter_html: true}))
+      Cdo::Markdown::Handler.register(Cdo::Markdown::Renderer, Cdo::Markdown::Renderer::OPTIONS)
+      Cdo::Markdown::Handler.register_html_safe(Redcarpet::Render::Safe)
+      Cdo::Markdown::Handler.register_inline(Redcarpet::Render::Inline.new({filter_html: true}))
     end
 
     def response
